moved bootstrapping code into respective daemon directories; simplified CMake configuration process
