--- src/sadmd/CMakeLists.txt
@@ -1,2 +1,0 @@
-# src/sadmd/CMakeLists.txt
-add_executable (sadmd sadmd.cpp)
